+++ configdirector-server-sdk/CHANGELOG.md
@@ -8,6 +8,12 @@ follows [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [1.4.0] - 2026-09-21
+
+### Changed
+
+- Allow consuming SDKs, like the OpenFeature provider, to adjust the SDK identity used to identify against the SDK server.
+
 ## [1.3.0] - 2026-09-20
 
 ### Fixed